requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
supabase==1.0.0
pydantic==1.10.7
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
flask-caching==1.10.1
psycopg2-binary==2.9.6
uvicorn[standard]==0.22.0
astral==3.2
email-validator==2.2.0
